Upload Methods CreateContact

+++ GaHipHop_API/Migrations/MyDbContextModelSnapshot.cs
@@ -89,11 +89,11 @@ namespace GaHipHop_API.Migrations
 
             modelBuilder.Entity("GaHipHop_Repository.Entity.Contact", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    MySqlPropertyBuilderExtensions.UseMySqlIdentityColumn(b.Property<int>("Id"));
+                    MySqlPropertyBuilderExtensions.UseMySqlIdentityColumn(b.Property<long>("Id"));
 
                     b.Property<string>("Email")
                         .IsRequired()
